+++ example/lib/repositories/adyen_blik_component_repository.dart
@@ -1,0 +1,85 @@
+import 'package:adyen_checkout/adyen_checkout.dart';
+import 'package:adyen_checkout_example/config.dart';
+import 'package:adyen_checkout_example/repositories/adyen_base_repository.dart';
+
+class AdyenBlikComponentRepository extends AdyenBaseRepository {
+  AdyenBlikComponentRepository({
+    required super.service,
+  });
+
+  Future<SessionCheckout> createSessionCheckout(
+      BlikComponentConfiguration blikComponentConfiguration) async {
+    final sessionResponse = await _fetchSession();
+    final test = await AdyenCheckout.session.create(
+      sessionId: sessionResponse['id'],
+      sessionData: sessionResponse['sessionData'],
+      configuration: blikComponentConfiguration,
+    );
+    return test;
+  }
+
+  Future<Map<String, dynamic>> _fetchSession() async {
+    String returnUrl = await determineBaseReturnUrl();
+    returnUrl += '/adyenPayment';
+
+    final sessionRequestBody = <String, dynamic>{
+      'merchantAccount': Config.merchantAccount,
+      'amount': {
+        'currency': Config.amount.currency,
+        'value': Config.amount.value,
+      },
+      'returnUrl': returnUrl,
+      'reference':
+          'flutter-session-test_${DateTime.now().millisecondsSinceEpoch}',
+      'countryCode': Config.countryCode,
+      'shopperLocale': Config.shopperLocale,
+      'shopperReference': Config.shopperReference,
+      'channel': determineChannel(),
+      'storePaymentMethodMode': 'disabled',
+      'shopperInteraction': 'Ecommerce',
+    };
+
+    return await service.createSession(sessionRequestBody);
+  }
+
+  Future<Map<String, dynamic>> fetchPaymentMethods() async {
+    return await service.fetchPaymentMethods(<String, dynamic>{
+      'merchantAccount': Config.merchantAccount,
+      'countryCode': Config.countryCode,
+      'channel': determineChannel(),
+      'shopperReference': Config.shopperReference,
+    });
+  }
+
+  Future<PaymentEvent> onSubmit(
+    Map<String, dynamic> data, [
+    Map<String, dynamic>? extra,
+  ]) async {
+    String returnUrl = await determineBaseReturnUrl();
+    returnUrl += '/adyenPayment';
+
+    final paymentsRequestBody = <String, dynamic>{
+      'merchantAccount': Config.merchantAccount,
+      'shopperReference': Config.shopperReference,
+      'reference': 'flutter-test_${DateTime.now().millisecondsSinceEpoch}',
+      'returnUrl': returnUrl,
+      'amount': {
+        'value': Config.amount.value,
+        'currency': Config.amount.currency,
+      },
+      'countryCode': Config.countryCode,
+      'channel': determineChannel(),
+      'shopperInteraction': 'Ecommerce',
+    };
+
+    paymentsRequestBody.addAll(data);
+    final response = await service.postPayments(paymentsRequestBody);
+    return paymentEventHandler.handleResponse(jsonResponse: response);
+  }
+
+  Future<PaymentEvent> onAdditionalDetails(
+      Map<String, dynamic> additionalDetails) async {
+    final response = await service.postPaymentsDetails(additionalDetails);
+    return paymentEventHandler.handleResponse(jsonResponse: response);
+  }
+}